+++ app/frontend/stylesheets/chat_room.scss
@@ -160,7 +160,7 @@
   position: relative;
   .chatroombox{
     position: fixed;
-    top: 310px;
+    bottom: 0px;
     right: 300px;
   }
 }
