+++ app/frontend/stylesheets/index.js
@@ -5,3 +5,4 @@ import "tailwindcss/components";
 import "tailwindcss/utilities";
 import "./chat_room.scss";
 import "./item_table.css"
+import "./posts_block.css"
