+++ app/frontend/stylesheets/posts_block.css
@@ -1,0 +1,36 @@
+#added-post-sample{
+  display: none;
+}
+.posts-block{
+  width: 0;
+  height: 0;
+  overflow: hidden;
+}
+.is-active-for-posts-block{
+  height: 100vh;
+  width: 300px;
+  position: absolute;
+  right: 0;
+  overflow-y: auto;
+  background-color: white;
+}
+.posts-header{
+  padding: 10px 15px;
+  font-size: 20px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  background-color: #f5f5f5;
+}
+.item-name-for-posts-block{
+  width: 100%;
+  height: 25px;
+  padding: 5px;
+  overflow: hidden;
+  white-space:nowrap;
+  text-overflow:ellipsis;
+  text-align: center;
+}
+.posts-body{
+  padding: 0 15px;
+}
