+++ app/views/items/posts.js.erb
@@ -1,0 +1,1 @@
+$('#posts-block').html("<%= j render('items/partials/posts') %>")
