+++ app/views/items/可刪index.html.erb
