+++ app/views/posts/create.js.erb
@@ -1,0 +1,2 @@
+$('#posts-show').prepend("<%= j render('posts/partials/createPost') %>");
+$('#posts-content').val("")
