+++ app/views/posts/index.html.erb
@@ -1,0 +1,1 @@
+<p> hi 我是post的index，你不應該看到我</p>
