--- app/views/shared/frontend/_post_modal.html.erb
@@ -1,36 +1,0 @@
-<div class="modal-post-block" id="modal-posts-content">
-  <div class="modal-card-post" >
-    <header class="modal-card-head">
-      <p class="modal-card-title post-card-title"></p>
-      <button class="delete btn-cancel-modal" aria-label="close"></button>
-    </header>
-
-    <section class="modal-card-body">
-      <div class="card adding-post">
-        <%= form_with(model: @post, local: true) do |f|%>
-          <%= f.label :postcontent, "留言：", class:"text-2xl p-2"%>
-          <%= f.text_area :postcontent, id:"posts-content", class:"textarea border border-dash border-green-300 rounded-md px-3 py-3", placeholder:"請在此輸入文字..." %>
-          <%= f.submit '送出', class:"button is-success btn-save-adding my-2", id:"btn-post-submit" %>
-        <% end %>
-      </div>
-      <% @posts.each do |post|%>
-        <%# if  == post.item_id %>
-          <div class="card added-post mb-3" data-item-id="<%= post.item_id %>" >
-            <div class="card-content">
-              <span class="author-email bg-yellow-200 p-1"><%= User.find(post.user_id).email %></span>
-              <p class="content mt-2"><%= post.content %></p>
-            </div>
-            <div class="card-footer">
-              <button class="card-footer-item">
-                <span class="far fa-thumbs-up"> Like</span>
-              </button>
-              <button class="card-footer-item">
-                <span class="fas fa-reply"> Reply</span>
-              </button>
-            </div>
-          </div>
-        <%# end %>
-      <% end %>
-    </section>
-  </div>  
-</div>
